+++ CHANGES.rst
@@ -7,6 +7,7 @@ development
 - Fix ``nearest-station`` with ``--format=json``
 - Fix filtering by ``int64``-type identifiers, see GH-7
 - Fix SQL filtering with DuckDB
+- Tests: Add software tests
 
 2020-12-29 0.11.0
 =================